requirements.txt
click
click-command-tree
click-help-colors
click-plugins
python-gitlab
psutil
requests
PyYAML